// src/objects/enums/ItemId.ts
export enum ItemId {
  I20001000 = '20001000',
  I20001001 = '20001001',
  I20001002 = '20001002',
  I20001003 = '20001003',
  I20001004 = '20001004',
  I20001005 = '20001005',
  I20001006 = '20001006',
  I20001007 = '20001007',

  I20002000 = '20002000',
  I20002001 = '20002001',
  I20002002 = '20002002',
  I20002003 = '20002003',
  I20002004 = '20002004',
  I20002005 = '20002005',
  I20002006 = '20002006',
  I20002007 = '20002007',

  I20003000 = '20003000',
  I20003001 = '20003001',
  I20003002 = '20003002',
  I20003003 = '20003003',
  I20003004 = '20003004',
  I20003005 = '20003005',
  I20003006 = '20003006',
  I20003007 = '20003007',

  I20004001 = '20004001',
  I20004002 = '20004002',
  I20004003 = '20004003',
  I20004004 = '20004004',
  I20004005 = '20004005',
  I20004006 = '20004006',
  I20004007 = '20004007',
  I20004008 = '20004008',
  I20004009 = '20004009',
  I20004010 = '20004010',
  I20004011 = '20004011',
  I20004012 = '20004012',
  I20004013 = '20004013',
  I20004014 = '20004014',
}